Replaced the old version of the Grid Operator API script with six new scripts that focus on individual features, such as Rest/Redis connection and type of grid fees strategy implemented, for example. More specifically, the major changes are: - Removed pandas/openpyxl dependencies; - Adopted best practices used for backend code (used both Black and Pylint to check the code); - Added logs externally to the script; - Generally improved script's structure and logic.
