+++ gsy_e_sdk/setups/Grid_Operator_API_scripts/redis_grid_fees.py
@@ -1,12 +1,11 @@
 # flake8: noqa
+# pylint: disable=duplicate-code
+
 """
 Template file for markets management through the gsy-e-sdk api client
 """
 import os
-import csv
 from time import sleep
-from pendulum import from_format
-from gsy_framework.constants_limits import DATE_TIME_FORMAT, TIME_FORMAT_SECONDS
 from gsy_e_sdk.types import aggregator_client_type
 from gsy_e_sdk.redis_market import RedisMarketClient
 from gsy_e_sdk.utils import log_grid_fees_information
@@ -23,19 +22,17 @@ SLOT_LENGTH = 15  # leave as is
 
 class Oracle(aggregator_client_type):
     """Class to represent the Grid Operator client type."""
+
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.is_finished = False
 
-    def set_new_market_fee(self, market_info):
+    def set_new_market_fee(self):
         """Return the market fees for each market for the next time slot."""
         next_market_fee = {}
-        market_time = from_format(market_info["market_slot"], DATE_TIME_FORMAT)
         for area_uuid, area_dict in self.latest_grid_tree_flat.items():
             if area_dict["area_name"] in market_names:
-                next_market_fee[
-                    area_dict["area_name"]
-                ] = calculate_next_slot_market_fee(market_time, area_dict["area_name"])
+                next_market_fee[area_dict["area_name"]] = 10
 
                 self.add_to_batch_commands.grid_fees(
                     area_uuid=area_uuid,
@@ -51,8 +48,7 @@ class Oracle(aggregator_client_type):
                 current_market_fee[area_dict["area_name"]] = area_dict[
                     "current_market_fee"
                 ]
-        self.execute_batch_commands()
-        next_market_fee = self.set_new_market_fee(market_info)
+        next_market_fee = self.set_new_market_fee()
         self.execute_batch_commands()
         log_grid_fees_information(market_names, current_market_fee, next_market_fee)
 
@@ -63,41 +59,8 @@ class Oracle(aggregator_client_type):
         self.is_finished = True
 
 
-def read_fee_strategy():
-    "Return a dictionary containing the Time of Use strategy loaded from the CSV input file."
-    with open(
-        os.path.join(current_dir, "resources/ToU.csv"), newline="", encoding="utf-8"
-    ) as csvfile:
-        csv_rows = csv.reader(csvfile, delimiter=" ", quotechar="|")
-        headers = next(csv_rows)[0].split(";")
-        market_indexes = {}
-        planned_fee = {}
-        for market_name in market_names:
-            market_indexes.update({(market_name, headers.index(market_name))})
-        for row in csv_rows:
-            row = row[0].split(";")
-            for market in market_names:
-                planned_fee.update({(row[0], market): row[market_indexes[market]]})
-    return planned_fee
-
-
-def calculate_next_slot_market_fee(market_time, market_name):
-    """Return the market fee for the next time slot."""
-    slot_time = (
-        market_time.add(minutes=1 * SLOT_LENGTH).time().format(TIME_FORMAT_SECONDS)
-    )
-    if (slot_time, market_name) in fee_strategy:
-        next_fee = fee_strategy[(slot_time, market_name)]
-        if not isinstance(next_fee, (int, float)):
-            next_fee = float(next_fee.replace(",", "."))
-    else:
-        next_fee = None
-    return next_fee
-
-
 MarketClient = RedisMarketClient
 aggr = Oracle(aggregator_name=ORACLE_NAME)
-fee_strategy = read_fee_strategy()
 
 print()
 print("Connecting to markets ...")
